refactor(go-template): extract block-body/object memo value computation into memo/memo-value.ts
Move resolveBlockBodyMemoModuleConst and computeObjectMemoInitialValue
out of the adapter class into pure free functions (Roadmap B). They read
only state.localConstants / state.searchParamsLocals and parseLiteralExpression,
delegating value lowering to lowerCtorExpr — no new GoEmitContext member.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
